scripts/nfs-export-export-updater: start own unfsd server instance
So far, nfs-export-updater required to run 'runqemu-export-rootfs'
before.

This requirement is now replaced by automatically invoking the script.

Beside the obvious advantage of one command less to type, the main
reason for doing so is that this allows to auto-determine a proper
NFS_INSTANCE, NFSD_PORT, and MOUNTD_PORT.

The code for this is copied and adapted from the runqemu script.

With this change, one can now run multiple independent unfsd instances
on the same system.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>