scripts/nfs-export-updater: Support creating NFS export directory
If the export dir is not given, we no attempt to create it using the
'runqemu-extract-sdk' script.

Since this requires us to know the rootfs archive path in the deploy
folder, we need to guess it from the provided rootfs image name.

This is mainly what create_nfs_exportdir() does.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>