scripts/nfs-export-updater: robust nfs instance detection
This fixes a bug were a socket was already bound but nobody was
listening. So the connect would fail, marking the port as free, but the
later binding of the same port by unfsd would fail with 'Adress already
in use'.

As unfsd by default binds a udp and tcp port to all interfaces extend
the check to cover this.

Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>